Generalize Cosmos SendPacket and WriteAck event implementation
