+++ crates/cosmos/cosmos-relayer/src/chain/impls/mod.rs
@@ -1,4 +1,3 @@
-pub mod events;
 pub mod ext;
 pub mod fields;
 pub mod types;
